refactor: move EOSEAccountFast to commons, KMP-purified (Phase 2 prep)
EOSEAccountFast is used by the user/event finder assemblers that will move to
commonMain. Relocate it to commons.relays with KmpLock instead of
synchronized(...) so it passes the iOS verifyKmpPurity gate. The old
amethyst.service.relays location keeps a typealias, so its 7 existing importers
are untouched. (SincePerRelayMap/MutableTime/EOSERelayList were already commons
typealiases.)
